+++ pyams_lib/__init__.py
@@ -1,9 +1,0 @@
-from .cad import *
-from .electrical import *
-from .progressbar import *
-from .PyAMS import *
-from .std import *
-from .struct_ import *
-from .unewton import *
-from .utils import *
-